+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 -   Fix a bug in the deprecated callback handlers of Form component. #7356
+-   Add `hidden` legend position to `Chart`. #7378
 
 # 8.0.0