+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 8.1.0
+
 -   Fix a bug in the deprecated callback handlers of Form component. #7356
 -   Fix a bug in the `<DateFilter>` component where values were retained when switching between rules #7423
 -   Add `hidden` legend position to `Chart`. #7378